+++ webpack.config.js
@@ -53,8 +53,8 @@ module.exports = [
                 BUNDLED: true,
                 VERSION: `'${version}'`
             }),
-            // new webpack.optimize.ModuleConcatenationPlugin(),
-            // new uglify
+            new webpack.optimize.ModuleConcatenationPlugin(),
+            new uglify
         ]
     },